hanna-package {hanna}	R Documentation

Package hanna - simulations for winner-looser effects

Description

This package contains the code for the paper "Hermanussen et. al. (2023): Winner-loser effects improve social network efficiency among competitors with equal resource holding power."

The following objects / modules are available:

simul R code for the simulation of the effect

hgraph R utility code functions for working with graphs

Author(s)

Detlef Groth, University of Potsdam, Germany

[Package hanna version 0.1 Index]