hgraph {hanna}	R Documentation

Environment object with utility functions to work with graphs.

Description

The functions of this environment contain a few utility functions to work with graphs represented as adjacency matrix.

Methods

The following functions are implemented:

hgraph$average_path_length - calculate the average path length for a given adjacency matrix

hgraph$colors - create a color vector based on node degrees

hgraph$d2u - convert a directed into an undirected graph

hgraph$degree - calculate the degree centrality for all nodes

hgraph$eccentricity - calculate eccentricity centrality for all nodes

hgraph$graph - create some example graphs

hgraph$layout - calculate a plot layout for an adjacency matrix

hgraph$plot - plot an adjacency matrix

hgraph$shortest_paths - calculate the shortest paths between all nodes of an adjacency matrix

hgraph$triads - calculate the number of two and tri edge triads

Examples

set.seed(124) 
A=matrix(c(0,0,1,0,0,0, 
           0,0,1,0,0,9, 
           0,0,0,1,0,0, 
           0,0,0,0,1,1, 
           0,0,0,0,0,1, 
           0,0,0,0,0,0),ncol=6,byrow=TRUE) 
colnames(A)=rownames(A)=LETTERS[1:6] 
A 
hgraph$shortest_paths(A) 
hgraph$shortest_paths(A,mode="undirected") 

[Package hanna version 0.1 Index]