hgraph$average_path_length {hanna}	R Documentation

Calculate the average path length between all nodes of a graph.

Description

This function calculates the average shortest path length between all pair of nodes. In unconnected graphs it returns Inf, if the connected option is set to TRUE it returns the average path length for the connected nodes, you can as well give a value which should be added instead of Inf values, usually this is the number of nodes.

Usage

`hgraph$average_path_length(x,mode="directed",unconnected=FALSE,infinite=NULL)`

Arguments

x
adjacency matrix

mode
should be either "directed" or "undirected", default: "directed"

unconnected
should only unconnected nodes be used in calculation, default: FALSE

infinite
a value which should replace Inf values, usually it is the number of nodes, default: NULL

Value

matrix with the pairwise path lengths

Examples

set.seed(123) 
x=1 

[Package hanna version 0.1 Index]