hgraph$d2u {hanna}	R Documentation

Covert a directed graph into an undirected one.

Description

This function converts an directed graph into an undirected one, the resulting adjacency matrix will be symmetric.

Usage

`hgraph$d2u(x)`

Arguments

x
adjacency matrix

Value

adjacency matrix

Examples

A=hgraph$graph(type="werner") 
A 
U=hgraph$d2u(A) 
par(mfrow=c(1,2)) 
lay=hgraph$layout(A) 
plot(A,layout=lay) 
plot(U,layout=lay) 

[Package hanna version 0.1 Index]