hgraph$degree {hanna}	R Documentation

Calculate the degree centrality for all nodes.

Description

This function calculates the number of edges connected to a a node. For directed graph you can distinguish between in and outgoing nodes using the mode argument.

Usage

`hgraph$degree(x,mode="all")`

Arguments

x
adjacency matrix

mode
character string, either 'all' for 'in' and 'out' going edges or one of the latter, for undirected graphs the mode is always 'all', default: 'all'

Value

numeric vector with the number of connected edges

Examples

set.seed(124) 
A=hgraph$graph(type="werner") 
hgraph$degree(A) 
hgraph$degree(A,mode='in') 
hgraph$degree(A,mode='out') 

[Package hanna version 0.1 Index]