hgraph$eccentricity {hanna}	R Documentation

Calculate the eccentricity centrality for all nodes.

Description

This function calculates the inverse of the longest shortest path for each node. In unconnected graphs it returns Inf, if the connected option is set to TRUE it returns the value between connected nodes only, you can as well give a value which should be added instead of Inf values using the infinite argument, usually this is the number of nodes.

Usage

`hgraph$eccentricity(x,mode="directed",unconnected=FALSE,infinite=NULL)`

Arguments

x
adjacency matrix

mode
should be either "directed" or "undirected", default: "directed"

unconnected
should only unconnected nodes be used in calculation, default: FALSE

infinite
a value which should replace Inf values, usually it is the number of nodes,default: NULL

Value

matrix with the pairwise path lengths

Examples

set.seed(123) 
A=hgraph$graph() 
A 
hgraph$eccentricity(A) 
hgraph$eccentricity(A,mode="undireced") 

[Package hanna version 0.1 Index]