hgraph$graph {hanna}	R Documentation

Create a few example graphs.

Description

This function creates a new graph such as random, kite or werner graphs.

Usage

`hgraph$graph(type="werner",nodes=10,edges=15)`

Arguments

type
eithern a adjacency matrix or a graph type, either 'werner' or 'kite', default: 'werner'

nodes
for a random graph the number of edges, default: 10

edges
number of edges in the graph, default: 15

Value

adjacency matrix

Examples

A = hgraph$graph() 
A 

[Package hanna version 0.1 Index]