hgraph$shortest_paths {hanna}	R Documentation

Calculate the shortest path between all nodes of a graph.

Description

This function calculates the shortest path between all pair of nodes. In unconnected graphs it returns Inf, if the connected option is set to TRUE it returns the average path length for the connected nodes, you can as well give a value which should be added instead of Inf values, usually this is the number of nodes.

Usage

`hgraph$shortest_paths(x,mode="directed",weighted=FALSE,FUN=mean)`

Arguments

x
adjacency matrix

mode
should be either "directed" or "undirected", default: "directed"

weighted
should be taken as weighted graph, default: FALSE

FUN
for directed graphs with multiedges how to compute the entry in undirected graphs

Value

matrix with the pairwise path lengths

Examples

set.seed(123) 
A=matrix(rbinom(25,1,p=0.3),ncol=5) 
colnames(A)=rownames(A)=LETTERS[1:5] 
A[lower.tri(A)]=t(A)[lower.tri(A)] 
A 
hgraph$shortest_paths(A,mode="undirected") 

[Package hanna version 0.1 Index]