hgraph$triads {hanna}	R Documentation

Calculate the number of two and tri edge triads.

Description

This function calculates the number of two and tri edge triads for directed graphs.

The following triads are possible:

double-dominant (dd) A->B; A->C

double-subordinate (ds) A->B; C->B

pass-along (pa) A->B->C

transitive (tr) A->B; A->C ; B->C

cycle (cy) A->B->C->A

Usage

`hgraph$triads(x,percent=FALSE)`

Arguments

x
adjacency matrix

percent
should the results be return not wihh total numbers but in percent, default: FALSE

Value

list object with the following components:

ddnumber of double dominant triads

dsnumber of double subordinate triads

panumber of pass-along triads

trnumber of transitive triads

cynumber of cycle triads

Examples

A = matrix(rbinom(49,1,p=0.3),nrow=7) 
diag(A)=0 
A[lower.tri(A)]=0 
rownames(A)=colnames(A)=LETTERS[1:7] 
hgraph$plot(A) 
unlist(hgraph$triads(A)) 

[Package hanna version 0.1 Index]