simul {hanna}	R Documentation

Environment obkject with functions do simulate winner-looser effects.

Description

The functions within the simul environment perform simulations of winner-looser effects for the paper ...

Methods

simul$compare - compare the different models for a certain number of seasons

simul$graph - create a adjacency matrix out of the results for a match season

simul$pairings - create round pairings for a season

simul$season - create matches for everyone against everyone using the given model

Examples

set.seed(124) 
res=simul$season(LETTERS[1:6],model="null")  
res 
hgraph$plot(res$M) 

[Package hanna version 0.1 Index]