simul$compare {hanna}	R Documentation

Compare the different models for a certain number of seasons.

Description

This function does a comparison for different models determine the amount of triads after a certain number of seasons.

Usage

 `simul$compare(n=5,agents=12,seasons=3)` 

Arguments

n
how many repeats per model, default: 5

agents
how many agents/teams, default: 12

seasons
ow many seasons, default: 3

Value

data frame with the results, last column model type

Examples

 set.seed(128) 
 par(mfrow=c(1,3))  
 res.df=simul$compare(n=5,seasons=3) 
 for (mod in c("null","chance","gain")) {  
   rest=t(scale(t(res.df[res.df$model==mod,1:5]))) 
   boxplot(rest,main=mod,ylim=c(-2,2))  
   lines(1:5,apply(rest,2,median)) 
 } 

[Package hanna version 0.1 Index]