simul$graph {hanna}	R Documentation

Create a adjacency matrix out of the results for a match season.

Description

This function creates an adjacency matrix for an undirected or a directed graph out of the results of season where every team played against every other team. The edges will be directed from the winning to the loosing team in case the mode is "win" or between drawing teams in case the mode is "draw".

Usage

 `simul$graph(x,mode="draw")` 

Arguments

x
a season matrix with wins encoded as 1, losses as -1 and draws as 0

mode
either draw or win, default: 'draw'

Value

Adjacency matrix

Examples

set.seed(123) 
res=simul$season(LETTERS[1:6],model="null")  
res$M 
U = simul$graph(res$M,mode='draw') 
U 
D = simul$graph(res$M,mode='win') 
U 

[Package hanna version 0.1 Index]