simul$pairings {hanna}	R Documentation

Create matching pairs everyone against every one.

Description

This function creates pairings for a tournament where in every round item, teams, etc get new partners for playing.

Usage

 `simul$pairings(x)` 

Arguments

x
symmetric matrix with row and colum names being the same, or a vector of names

Details

This function allows you to create pairings between different teams or players for a tournament.

Value

data frame with the pairings for each round

Examples

mt=matrix(0,nrow=4,ncol=4) 
rownames(mt)=colnames(mt)=LETTERS[1:4] 
simul$pairings(mt) 
simul$pairings(c("ABA","CDE","EFG")) 

[Package hanna version 0.1 Index]