simul$season {hanna}	R Documentation

Create matches for everyone against everyone using the given model.

Description

This function creates pairings for a tournament where in every round. The actual match will give chances based on a certain amount of tokens in dependence of the given model.

Usage

 `simul$season(x,token=rep(length(x),length(x)),model="null", 
                           min.value=4,memory=NULL,memory.length=1)` 

Arguments

x
vector of teams

token
vector of token for each team, which might influence the match outcone, depending on the given model, defaults: 5

model
the model given as string, possible values are 'null', 'chance','gain' or 'last', default: 'null'

min.value
for the model 'last' how low is the minimal value for each team, default: 4

memory
optional vector of last results for each agent

memory.length
how many of last results should be stored, default: 1

Details

This function allows you to create matches for all against all in a season and performs the matches based on the given model and the current number of tokens for each team.

Value

list with three components:

Mmatrix of results where 1 is a win, -1 is a loss and 0 is a draw

tokenvector of current tokens for each team

modelthe choosen model

memorylist with last results for each agent

Examples

set.seed(123) 
res=simul$season(LETTERS[1:6],model="null")  
res 

[Package hanna version 0.1 Index]